Move to a single provider
