--- Example/Backpack/Utils/FeatureStories/Groups/PriceGroupsSwiftUI.swift
@@ -1,43 +1,0 @@
-/*
- * Backpack - Skyscanner's Design System
- *
- * Copyright 2018 Skyscanner Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import SwiftUI
-import Backpack
-
-struct PriceGroupsSwiftUIProvider {
-    let showPresentable: (Presentable) -> Void
-    
-    private func presentablePrice<Content: View>(
-        _ title: String,
-        view: Content
-    ) -> CellDataSource {
-        PresentableCellDataSource.custom(
-            title: title,
-            customController: { ContentUIHostingController(view) },
-            showPresentable: showPresentable
-        )
-    }
-        
-    func swiftUIGroups() -> [Components.Group] {
-        SingleGroupProvider(
-            cellDataSources: [
-                presentablePrice("Small", view: PriceExampleView(size: .small)),
-                presentablePrice("Large", view: PriceExampleView(size: .large))
-            ]
-        ).groups()
-    }
-}
